refactor script overview page
