+++ apps/src/code-studio/progress.js
@@ -1,18 +1,12 @@
 import $ from 'jquery';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {Provider} from 'react-redux';
 import _ from 'lodash';
-import queryString from 'query-string';
 import clientState from './clientState';
-import {convertAssignmentVersionShapeFromServer} from '@cdo/apps/templates/teacherDashboard/shapes';
-import UnitOverview from './components/progress/UnitOverview.jsx';
 import DisabledBubblesModal from './DisabledBubblesModal';
 import DisabledBubblesAlert from './DisabledBubblesAlert';
 import {getStore} from './redux';
-import {registerReducers} from '@cdo/apps/redux';
-import {setViewType, ViewType} from './viewAsRedux';
-import {getHiddenLessons, initializeHiddenScripts} from './hiddenLessonRedux';
+import {getHiddenLessons} from './hiddenLessonRedux';
 import {TestResults} from '@cdo/apps/constants';
 import {
   initProgress,
@@ -26,15 +20,7 @@ import {
   useDbProgress
 } from './progressRedux';
 import {setVerified} from '@cdo/apps/code-studio/verifiedInstructorRedux';
-import {
-  selectSection,
-  setSections,
-  setPageType,
-  pageTypes
-} from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
-import googlePlatformApi, {
-  loadGooglePlatformApi
-} from '@cdo/apps/templates/progress/googlePlatformApiRedux';
+import {pageTypes} from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
 import {renderTeacherPanel} from './teacherPanelHelpers';
 
 var progress = module.exports;
@@ -243,70 +229,15 @@ function extractLevelResults(userProgressResponse) {
  * @param {boolean} scriptData.age_13_required
  * Render our progress on the course overview page.
  */
-progress.renderCourseProgress = function(scriptData) {
+progress.initCourseProgress = function(scriptData) {
   const store = getStore();
   initializeStoreWithProgress(store, scriptData, null, true);
-  initializeStoreWithSections(store, scriptData);
-  if (scriptData.user_type === 'teacher') {
-    initializeGooglePlatformApi(store);
-  }
 
   if (scriptData.student_detail_progress_view) {
     store.dispatch(setStudentDefaultsSummaryView(false));
   }
-  progress.initViewAs(store, scriptData.user_type);
+
   queryUserProgress(store, scriptData, null);
-
-  const teacherResources = (scriptData.teacher_resources || []).map(
-    ([type, link]) => ({
-      type,
-      link
-    })
-  );
-
-  store.dispatch(initializeHiddenScripts(scriptData.section_hidden_unit_info));
-
-  store.dispatch(setPageType(pageTypes.scriptOverview));
-
-  const mountPoint = document.createElement('div');
-  $('.user-stats-block').prepend(mountPoint);
-
-  ReactDOM.render(
-    <Provider store={store}>
-      <UnitOverview
-        id={scriptData.id}
-        courseId={scriptData.course_id}
-        courseTitle={scriptData.course_title}
-        courseLink={scriptData.course_link}
-        excludeCsfColumnInLegend={!scriptData.csf}
-        teacherResources={teacherResources}
-        migratedTeacherResources={scriptData.migrated_teacher_resources}
-        studentResources={scriptData.student_resources || []}
-        showCourseUnitVersionWarning={
-          scriptData.show_course_unit_version_warning
-        }
-        showScriptVersionWarning={scriptData.show_script_version_warning}
-        showRedirectWarning={scriptData.show_redirect_warning}
-        redirectScriptUrl={scriptData.redirect_script_url}
-        versions={convertAssignmentVersionShapeFromServer(scriptData.versions)}
-        courseName={scriptData.course_name}
-        showAssignButton={scriptData.show_assign_button}
-        userId={scriptData.user_id}
-        assignedSectionId={scriptData.assigned_section_id}
-        showCalendar={scriptData.showCalendar}
-        weeklyInstructionalMinutes={scriptData.weeklyInstructionalMinutes}
-        unitCalendarLessons={scriptData.calendarLessons}
-        isMigrated={scriptData.is_migrated}
-        scriptOverviewPdfUrl={scriptData.scriptOverviewPdfUrl}
-        scriptResourcesPdfUrl={scriptData.scriptResourcesPdfUrl}
-        showUnversionedRedirectWarning={
-          scriptData.show_unversioned_redirect_warning
-        }
-        isCsdOrCsp={scriptData.isCsd || scriptData.isCsp}
-      />
-    </Provider>,
-    mountPoint
-  );
 };
 
 progress.retrieveProgress = function(scriptName, scriptData, currentLevelId) {
@@ -315,27 +246,6 @@ progress.retrieveProgress = function(scriptName, scriptData, currentLevelId) {
     initializeStoreWithProgress(store, scriptData, currentLevelId, true);
     queryUserProgress(store, scriptData, currentLevelId);
   });
-};
-
-/* Set our initial view type (Participant or Instructor) from current user's user_type
- * or our query string. */
-progress.initViewAs = function(store, userType) {
-  // Default to Participant, unless current user is a teacher
-  let initialViewAs = ViewType.Participant;
-  if (userType === 'teacher') {
-    //TODO(dmcavoy): Update to check instructor
-    initialViewAs = ViewType.Instructor;
-  }
-
-  // If current user is not a student (ie, a teacher or signed out), allow the
-  // 'viewAs' query parameter to override;
-  if (userType !== 'student') {
-    //TODO(dmcavoy): Update to check participant
-    const query = queryString.parse(location.search);
-    initialViewAs = query.viewAs || initialViewAs;
-  }
-
-  store.dispatch(setViewType(initialViewAs));
 };
 
 /**
@@ -438,35 +348,4 @@ function initializeStoreWithProgress(
   }
 
   store.dispatch(setIsAge13Required(scriptData.age_13_required));
-}
-
-function initializeStoreWithSections(store, scriptData) {
-  const sections = scriptData.sections;
-  if (!sections) {
-    return;
-  }
-
-  const currentSection = scriptData.section;
-  if (!currentSection) {
-    // If we don't have a selected section, simply set sections and we're done.
-    store.dispatch(setSections(sections));
-    return;
-  }
-
-  // If we do have a selected section, merge it with the minimal data in the
-  // `sections` array before storing in redux.
-  const idx = sections.findIndex(section => section.id === currentSection.id);
-  if (idx >= 0) {
-    sections[idx] = {
-      ...sections[idx],
-      ...currentSection
-    };
-  }
-  store.dispatch(setSections(sections));
-  store.dispatch(selectSection(currentSection.id.toString()));
-}
-
-function initializeGooglePlatformApi(store) {
-  registerReducers({googlePlatformApi});
-  store.dispatch(loadGooglePlatformApi()).catch(e => console.warn(e));
 }
